requirements.txt
Django==1.11.4
django-model-utils==3.0.0
django-notifications-hq==1.3
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
geopy==1.11.0
ipaddress==1.0.18
jsonfield==2.0.2
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
